fix compile errors

+++ lib/pipeline/nodes/CreatePointCloudNode.cpp
@@ -151,8 +151,8 @@ namespace pointmapper::pipeline {
         pc.axisScale.w = fd->axisScale[3];
         pc.depthScale = fd->depthUnits;
 
-        int groupsX = std::max(1, p.width / 8);
-        int groupsY = std::max(1, p.height / 8);
+        int groupsX = std::max(1, static_cast<int>(p.width) / 8);
+        int groupsY = std::max(1, static_cast<int>(p.height) / 8);
 
         cloudPipeline->DispatchWorkgroups(bundle.encoder, groupsX, groupsY, 1, &pc);
     }
